+++ Editor/MultiSceneToolsMenuItems.cs
@@ -21,13 +21,13 @@ namespace HH.MultiSceneTools
 {
     public static class MultiSceneToolsMenuItems
     {
-        [MenuItem("Multi Scene Tools/Reload project Collections")]
+        [MenuItem("Tools/Multi Scene Tools/Reload project Collections")]
         static void UpdateCollections()
         {
             MultiSceneToolsConfig.instance.UpdateCollections();
         }
 
-        [MenuItem("Multi Scene Tools/Changelog")]
+        [MenuItem("Tools/Multi Scene Tools/Changelog")]
         public static void OpenChangelog()
         {
             Application.OpenURL("https://github.com/HenrysHouses/MultiSceneTools/blob/main/CHANGELOG.md");